Create a pages dir inside tmp

+++ convert-pdf.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"regexp"
 
 	"github.com/couchbaselabs/logg"
 )
@@ -43,7 +44,11 @@ func (c ConvertPdf) preprocess(ocrRequest *OcrRequest) error {
 		tmpFileNameOutput,
 	)
 
-	tmpDir := createTempDir()
+	tmpDir, err_tmpdir := createTempDir()
+	if err_tmpdir != nil {
+		logg.LogFatal("Error running command: %s.", err_tmpdir)
+		return err_tmpdir
+	}
 	defer os.RemoveAll(tmpDir)
 
 	tmpDirFiles := fmt.Sprintf("%s/%s_%s.pdf", tmpDir, ocrRequest.Name, "%03d")
@@ -66,27 +71,31 @@ func (c ConvertPdf) preprocess(ocrRequest *OcrRequest) error {
 		"output",
 		tmpDirFiles,
 	).CombinedOutput()
-	if err != nil {
+	if err_pdftk != nil {
 		logg.LogFatal("Error running command: %s.  out: %s", err_pdftk, out_pdftk)
 	}
 	logg.LogTo("PREPROCESSOR_WORKER", "output: %v", string(out_pdftk))
 
-	err_walk := filepath.Walk(tmpDir, func(path string, f os.FileInfo, err error) error {
-		out_imagemagick, err_imagemagick := exec.Command(
-			"convert",
-			"-density",
-			"300",
-			"-depth",
-			"8",
-			"-alpha",
-			"Off",
-			path,
-			tmpFileNameOutput,
-		).CombinedOutput()
+	err_walk := filepath.Walk(tmpDir + "/", func(path string, f os.FileInfo, err error) error {
 		if err != nil {
-			logg.LogFatal("Error running command: %s.  out: %s", err_imagemagick, out_imagemagick)
+			logg.LogFatal("Error running command: %s.", err)
 		}
-		logg.LogTo("PREPROCESSOR_WORKER", "Path is: %s, Output: %v", path, string(out_imagemagick))
+		logg.LogTo("PREPROCESSOR_WORKER", "Path is: %s. Name is: %s.", path, f.Name())
+		matched, _ := regexp.MatchString("^.*?_[0-9]{3}\\.pdf", f.Name())
+		if matched {
+			out_imagemagick, err_imagemagick := exec.Command(
+				"convert",
+				"-density",
+				"300",
+				"-depth",
+				"8",
+				"-alpha",
+				"Off",
+				path,
+				tmpFileNameOutput,
+			).CombinedOutput()
+			logg.LogTo("PREPROCESSOR_WORKER", "output: %v, error: %v. ", string(out_imagemagick), err_imagemagick)
+		}
 		return nil
 	})
 	if err_walk != nil {
